+++ src/nice_int/nice_u64.rs
@@ -135,7 +135,7 @@ mod tests {
 		const SAMPLE_SIZE: usize = 1_000_000;
 
 		#[cfg(miri)]
-		const SAMPLE_SIZE: usize = 1000; // Miri runs way too slow for a million tests.
+		const SAMPLE_SIZE: usize = 500; // Miri runs way too slow for a million tests.
 
 		// Check the min and max.
 		assert_eq!(NiceU64::from(0_u64).as_str(), "0");